+++ BE/src/record/record.module.ts
@@ -4,10 +4,11 @@ import { RecordService } from './record.service';
 import { RedisModule } from '../config/redis/redis.module';
 import { UserDBManager } from '../config/query-database/user-db-manager.service';
 import { UsageModule } from '../usage/usage.module';
+import { RecordInitializeService } from './record-initialize.service';
 
 @Module({
   imports: [RedisModule, UsageModule],
   controllers: [RecordController],
-  providers: [RecordService, UserDBManager],
+  providers: [RecordService, UserDBManager, RecordInitializeService],
 })
 export class RecordModule {}
